minor ssh config ssh-notify tweak
Tried adding which ssh key was used, but failed.
man ssh_config says many other '%' tokens can be
used in the params to LocalCommand, but in practice
only %r and %h work, and the other keys don't seem
to include a description of the chosen key after all.
So give up on that.

Retain the probably-better use of  instead of ,
and a slightly different message wording.