Fix primary-monitor to sort monitors by horizontal position
This prevents the massive bug in the previous version whereby
for integer X, monitor X varied from run to run, as xrandr would
output them in a different order depending on which is primary.

We now sort them by horizontal position, giving a stable order,
and making it meaningful to make monitor '2' primary (ie. the
3rd (counting from 0) from the left).