Modernize nvim wrapper script 'ni'
Previously this unconditionally opened a new terminal to run nvim in.
This is disruptive: To use EDITOR=ni, special care has to be taken to
run the terminal app synchronously, and it prevents the same EDITOR
setting being shared in containers (since they can't open new terminal
GUI applications on the host desktop.)

So this new approach is simpler:
* Only open a new terminal if we are not already running in one
* Otherwise, use the one we're in
* If 'nvim' is not on the PATH, fallback to 'vim'

This allows us to share the EDITOR setting on the host and on containers
(simplifying .profile), remove the 'nis' alternate alias for 'ni', and
has bonuses such as allowing me to bind a desktop custom key shortcut to
'ni', and it just does the right thing (open a new terminal to run nvim
in it).

Also delete related wrapper 'niv', which I have not used in years.
Suspect it existed to quash noisy messages from the terminal or from
nvim, but I don't see any today, so perhaps this is no longer a problem.

Also: Install nvim in our lxc containers.