Stop 'cd' builtin printing dest. dir to stdout
This is puzzling. Yesterday, the 'cd' builtin started more frequently
outputting the destination directory to stdout after a successful
directory change, in scenarios when - I strongly believe - it used to be
silent.

It has always done this sometimes, for example, when a 'cd' command
specifies a directory that could only be found under the "~/src"
component of the CDPATH. This was expected.

'man bash' describes this behavior:

    If a non-empty directory name from CDPATH is used, or if - is the
    first argument, and the directory change is successful, the absolute
    pathname of the new working directory is written to the standard
    output.

This is exactly what I'm observing. From ~, the command `cd src/sca`,
which uses the non-empty initial '.' component of CDPATH, is outputting
the result to stdout. But I am certain this never used to happen on a
directory expressed as relative to '.'. What changed? Could it have
always been this way, and I just never noticed???

Changing CDPATH to just '~/src' to prevent this has problems. From ~,
'cd docs' goes to '~/src/docs' rather than '~/docs'. It is trying the
elements of CDPATH before falling back to a built-in fallback of '.'.

Hence the new setting, CDPATH=:~/src. The empty first component is
treated like '.', but does not output the destination to stdout.