Add aliases for 'git switch' and 'git restore'
These commands reify the two different uses of 'checkout'. If these are
the future, I might as well make them easy to use, to help untrain my
muscle memory from reaching for 'git commit'.

1. switch

Select a new current branch, eg:

Switches to a new branch which starts at existingcommit:

    git switch -c <newbranch> [<existingcommit>]

Switches to a commit in detached head:

    git switch --detach <commit>

2. restore

Change files to match their state at a given commit, eg:

By default, --worktree is implied, to update files, and the --source is HEAD,
so you can reset all files back to the last commit with:

    git restore .

Or restrict to particular files:

    git restore main.c

Or choose a different source to copy from:

    git restore --source HEAD~3 .

Or, instead of just affecting the files, put state into the index with
--staged. e.g. this will put an older version of main.c in the index
ready to be committed:

    git restore --staged --source HEAD~3 main.c